Extract build_cli lifecycle helper (mechanical)
Command: apply_patch move the constants and withBuildCliEnabled body into build_cli.dart, retaining a compatibility export.
